enable strong typing for MavenDeveloperPlugin
